fix(rt): isClosedForRead implies availableForRead is zero
The new roundTrip integrity checks found a case
where `isClosedForRead` flag was erroneously false. This was due
to the fact that the segments channel was drained but the current
segment had not been completely read. Updated the flag to explicitly
include the implied behavior that once `isClosedForRead` is true
the channel will never return any additional bytes from a read() call.