misc: bump smithy to 1.45.0
Also exclude a new protocol test which is failing in the sdk. The
test fails because it expects a blob members with a default of ""
to have the default populated when a value isn't explicitly provided,
but the sdk ignores defaults of "" for blob members. Fixing this is
non-trivial because using the default now would change the
nullability of members with this default (nullable -> non-nullable).
Keeping blobs with "" default nullable but using the default also
doesn't work because it means that migrating @required -> @default("")
would be a breaking change (granted, it still is now).